Added aie and aie_shim support to get_info function under xrt_device
