+++ src/runtime_src/core/common/api/xrt_device.cpp
@@ -30,6 +30,7 @@
 #include "core/common/info_memory.h"
 #include "core/common/info_platform.h"
 #include "core/common/query_requests.h"
+#include "core/common/info_aie.h"
 
 #include "handle.h"
 #include "native_profile.h"
@@ -287,6 +288,10 @@ get_info(info::device param, const xrt::detail::abi& abi) const
     return query::json_str(xrt_core::platform::pcie_info(handle.get()), abi);
   case info::device::dynamic_regions :         // std::string
     return query::json_str(xrt_core::memory::xclbin_info(handle.get()), abi);
+  case info::device::aie :         // std::string
+    return query::json_str(xrt_core::aie::aie_core(handle.get()), abi);
+  case info::device::aie_shim :         // std::string
+    return query::json_str(xrt_core::aie::aie_shim(handle.get()), abi);
   case info::device::host :                   // std::string
     boost::property_tree::ptree pt;
     xrt_core::get_xrt_build_info(pt);
